+++ duality/hub/hub.py
@@ -1,4 +1,37 @@
-from duality.misc.global_functions import *
+'''
+duality.hub CALLABLE FUNCTIONS:
+
+        .help() - shows available functions in the module.
+
+        .random_value(mean, st_dev, **rounded) - gives a random value of mean and standard deviation inputed, if rounded = 'y', value will be rounded.
+
+        .random_pool(mean, st_dev, pool_size, **rounded) - gives random values of mean and standard deviation inputed for the amount of values defined in the pool size, if rounded = 'y', values will be rounded.
+
+        .split_values(data, split_method) - splits the data using a split method character.
+
+        .join_values(data, join_method) - joins the data using a join metod character.
+
+        .replace_values(data, replaced_value, replacing_value) - replaces a defined value with a desired value.
+
+        .list_sort(data, array) - manually sorts data depending on defined array of indexes.
+
+        .index_sort(data, split_method, index_array) - sorts the indicies in a list of values based on the index array defined as [x,x,x].
+
+        .auto_sort(data, split_method, trigger = lambda x: x[0]) - automatically splits all values in a list and sorts them based on the added trigger as lambda x: [x[i], x[i]] and joins them back together.
+
+'''
+
+#metadata of the used package.
+from duality.misc._meta import (
+	__author__,
+	__copyright__,
+	__credits__,
+	__license__,
+	__version__,
+	__documentation__,
+	__contact__,
+	__donate__
+)
 
 #gives a random value of mean and standard deviation inputed, if rounded = 'y', value will be rounded.
 def random_value(mean, st_dev, **rounded):
@@ -84,16 +117,3 @@ def auto_sort(data, split_method, trigger = lambda x: x[0]):
                 d = '-'.join(i)
                 merged_final.append(d)
         return merged_final
-
-#shows detailed overview of available functions.
-def help():
-        print('duality.hub CALLABLE FUNCTIONS:\n')
-        print('.help() - shows available functions in the module.\n')
-        print('.random_value(mean, st_dev, **rounded) - gives a random value of mean and standard deviation inputed, if rounded = \'y\', value will be rounded.\n')
-        print('.random_pool(mean, st_dev, pool_size, **rounded) - gives random values of mean and standard deviation inputed for the amount of values defined in the pool size, if rounded = \'y\', values will be rounded.\n')
-        print('.split_values(data, split_method) - splits the data using a split method character.\n')
-        print('.join_values(data, join_method) - joins the data using a join metod character.\n')
-        print('.replace_values(data, replaced_value, replacing_value) - replaces a defined value with a desired value.\n')
-        print('.list_sort(data, array) - manually sorts data depending on defined array of indexes.\n')
-        print('.index_sort(data, split_method, index_array) - sorts the indicies in a list of values based on the index array defined as [x,x,x].\n')
-        print('.auto_sort(data, split_method, trigger = lambda x: x[0]) - automatically splits all values in a list and sorts them based on the added trigger as lambda x: [x[i], x[i]] and joins them back together.')