+++ duality/misc/_meta.py
@@ -9,5 +9,15 @@ __credits__ = [
 	'dkundih <https://github.com/dkundih>',
 
 ]
-__license__ = 'Apache 2.0 License'
-__version__ = '3.0.1'
+__license__ = 'Copyright (c) 2021- David Kundih. All rights reserved.\nLicensed under the Apache License, Version 2.0.\n\nFor more details about the license and terms of use visit the official duality documentation linked at https://github.com/dkundih/duality and https://pypi.org/project/duality'
+__version__ = '3.1.2'
+__documentation__ = 'https://github.com/dkundih/duality\nhttps://pypi.org/project/duality'
+__contact__ = 'dakundih@unin.hr\nkundihdavid@gmail.com'
+__donate__ = 'https://patreon.com/dkundih\nhttps://buymeacoffee.com/dkundih'
+
+#general setup confirmation.
+def main():
+    print('Setup successful.')
+
+if __name__ == '__main__':
+        main()